requirements.txt
africastalking==2.0.2
asgiref==3.11.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
cryptography==46.0.6
Django==6.0.3
django-daraja==1.3.0
djangorestframework==3.17.1
idna==3.11
pycparser==3.0
python-decouple==3.8
python-dotenv==1.2.2
PyYAML==6.0.3
requests==2.33.0
responses==0.26.0
schema==0.7.8
sqlparse==0.5.5
urllib3==2.6.3
